+++ src/main/java/com/example/atmbackend_ms/model/Transaction.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -22,6 +23,6 @@ public class Transaction {
     private String cardNumber;
     private LocalDateTime timestamp;
     private String type;
-    private double amount;
-    private double balanceAfterTransaction;
+    private BigDecimal amount;
+    private BigDecimal balanceAfterTransaction;
 }
